+++ test/example/script.js
@@ -1,10 +1,19 @@
-wiggle(document.getElementById("wiggle"), 0);
-const el = document.querySelector("#wiggle");
-// eslint-disable-next-line
-const dasd = document.getElementsByClassName("class");
+// exaples with selecting by a single id
+const idEl = document.getElementById("wiggle");
+idEl.id = "otherId";
 
-el.id = "wiggle";
-console.log(el, dasd);
+// exaples with selecting by a single class
+document.querySelector("#wiggle h2.header");
+document.querySelectorAll(".class p");
+
+// exaples with selecting by multiple selectors
+const classEl = document.getElementsByClassName("class");
+classEl.classList.add("otherClass");
+classEl.classList.remove("otherClass");
+classEl.classList.toggle("otherClass");
+classEl.className = "otherClass";
+
+wiggle(document.getElementById("wiggle"), 0);
 function wiggle(element, time) {
   element.style.marginLeft = Math.sin(time) * 10;
   time += 0.1;