+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-0.2.0:
+## 0.3.0:
+
+- improved the html processor: now it can run through internal js and css
+- improved the jsStrings processor: now it can run through querySelector and querySelectorAll methods + it adds selectors to the libraries on its own
+
+## 0.2.0:
 
 - added eslint and typescript
 - complete rewrite to typescript